[DSPDC-1753] Refresh credentials (#14)
Why

The initial take on refreshing google credentials for the jade API client does not work. Specifically, we are setting an API key variable, but the Jade client works in googleoauth mode, which ignores that setting and never calls the related refresh hook. Given that the client is auto-generated from an upstream OpenAPI python API generator, I am not optimistic about this being fixed on that end any time soon.

This PR

Sub-classes the Configuration class and hooks access to the access_token property. On get, we check if the underlying google credentials are valid; if not, we refresh and return a new access token
Moves the datarepo stuff to it's own package to avoid dumping the RefreshingConfiguration etc. into a single file or polluting the top-level resources package
Adds tests to verify that we can talk to Jade so this doesn't break again