[DSPDC-1798] Make job name a parameter of BeamRunner::run() (#21)
Why

We need to backport the beam runner changes to HCA. One wrinkle is that the K8S beam runner expects a job name; our initial approach was to pass that through as an arg in the arg dict. However, the local and dataflow beam runner implementations both automatically convert all args in the arg dict to scala direct runner args. Direct runner in turn errors out when it encounters the unexpected jobName variable.

This PR

Rather than adding special case code to the local and dataflow beam runners, I've opted to add job_name as an optional arg to the BeamRunner protocol's run method.