[DSPDC-1891] Flip load_incluster_config to true by default and externalize (#27)
Why

The load_incluster_config param was flipped to False while I tested locally, and I failed to flip it back to True. This will break downstream k8s runners.

This PR

Flips back to True by default, and allows the config to be set by downstream clients to avoid this situation in the future.